Changed basis rep to Int8.

+++ src/cga.jl
@@ -13,11 +13,11 @@ import Base.show
 
 
 #Basis
-const e1 = Odd{Float64}(Quaternion(0,1,0,0),Quaternion(0.0),Quaternion(0.0),Quaternion(0,-1,0,0))
-const e2 = Odd{Float64}(Quaternion(0,0,1,0),Quaternion(0.0),Quaternion(0.0),Quaternion(0,0,-1,0))
-const e3 = Odd{Float64}(Quaternion(0,0,0,1),Quaternion(0.0),Quaternion(0.0),Quaternion(0,0,0,-1))
-const e4 = Odd{Float64}(Quaternion(0.0),Quaternion(1,0,0,0),Quaternion(1,0,0,0),Quaternion(0.0))
-const f4 = Odd{Float64}(Quaternion(0.0),Quaternion(-1,0,0,0),Quaternion(1,0,0,0),Quaternion(0.0))
+const e1 = Odd{Int8}(Quaternion(0,1,0,0),Quaternion(0.0),Quaternion(0.0),Quaternion(0,-1,0,0))
+const e2 = Odd{Int8}(Quaternion(0,0,1,0),Quaternion(0.0),Quaternion(0.0),Quaternion(0,0,-1,0))
+const e3 = Odd{Int8}(Quaternion(0,0,0,1),Quaternion(0.0),Quaternion(0.0),Quaternion(0,0,0,-1))
+const e4 = Odd{Int8}(Quaternion(0.0),Quaternion(1,0,0,0),Quaternion(1,0,0,0),Quaternion(0.0))
+const f4 = Odd{Int8}(Quaternion(0.0),Quaternion(-1,0,0,0),Quaternion(1,0,0,0),Quaternion(0.0))
 const I5 = e1*e2*e3*e4*f4
 
 function basisCGA(T)
